==> python/06_prepare_population_data.py <==
"""
Script Name: 06_prepare_population_data.py
Purpose:
    Prepares external population data from Our World In Data (OWID) for use in
    the market penetration analysis. This script filters the dataset to the 
    most recent year (2023), selects relevant columns, standardises naming 
    conventions, and outputs a clean CSV containing population data only for 
    countries that appear in the gaming dataset.
Inputs:
    - data_external/population.csv
Outputs:
    - data_external/population_clean.csv

Author: Shian Raveneau-Wright

Notes:
    Data Source:
    Ritchie, H., et al. (2023). “Population.” Our World in Data.
    Retrieved from: https://ourworldindata.org/population
"""

import pandas as pd
from pathlib import Path

# Define paths
BASE_DIR = Path(".")
EXTERNAL_DIR = BASE_DIR / "data_external"
EXTERNAL_DIR.mkdir(exist_ok=True)

POP_CSV = EXTERNAL_DIR / "population.csv"
POP_CLEAN_CSV = EXTERNAL_DIR / "population_clean.csv"

# Load OWID population data
pop_df = pd.read_csv(POP_CSV)

# Inspect columns
print(pop_df.columns)

# Filter for year 2023 only
pop_2023 = pop_df[pop_df["Year"] == 2023]

# Select only necessary columns
pop_2023 = pop_2023[["Entity", "Population (historical)"]]

# Rename columns for consistency
pop_2023 = pop_2023.rename(columns={
    "Entity": "country",
    "Population (historical)": "population"
})

countries_in_data = [
"Italy",
"United States",
"United Kingdom",
"Spain",
"Russian Federation",
"Mexico",
"Canada",
"Poland",
"Australia",
"Brazil",
"Germany",
"Argentina",
"Bulgaria",
"Singapore",
"Portugal",
"Chile",
"Thailand",
"Japan",
"New Zealand",
"France",
"Hong Kong",
"Switzerland",
"Slovakia",
"Saudi Arabia",
"Belgium",
"Ireland",
"India",
"Netherlands",
"Malaysia",
"United Arab Emirates",
"South Africa",
"Greece",
"Croatia",
"Austria",
"Denmark",
"Finland",
"Indonesia",
"Czechia",
"Türkiye",
"Sweden",
"Taiwan, Province of China",
"Hungary",
"Israel",
"Iceland",
"Ukraine",
"Colombia",
"Kuwait",
"Malta",
"Lebanon",
"Norway",
"Qatar",
"Bahrain",
"Romania",
"Peru",
"Korea, Republic of",
"Oman",
"Luxembourg",
"Cyprus",
"Slovenia",
"China",
"El Salvador",
"Ecuador",
"Costa Rica",
"Honduras",
"Uruguay",
"Nicaragua",
"Guatemala",
"Paraguay",
"Bolivia, Plurinational State of",
"Panama",
"Korea, Democratic People's Republic of",
"Heard Island and McDonald Islands",
"Macao",
"Viet Nam",
"Liechtenstein",
"Somalia",
"Bosnia and Herzegovina",
"Philippines",
"Gibraltar",
"Nepal",
"Moldova, Republic of",
"Jamaica",
"Holy See (Vatican City State)",
"Palestine, State of",
"Congo",
"Libya",
"Afghanistan",
"Syrian Arab Republic",
"Northern Mariana Islands",
"Saint Helena, Ascension and Tristan da Cunha",
"Egypt",
"Venezuela, Bolivarian Republic of",
"Saint Vincent and the Grenadines",
"Uzbekistan",
"Brunei Darussalam",
"Bahamas",
"Virgin Islands, U.S.",
"Guernsey",
"United States Minor Outlying Islands",
"Iraq",
"Serbia",
"Senegal",
"Christmas Island",
"Mongolia",
"Sierra Leone",
"Turks and Caicos Islands",
"Fiji",
"Yemen",
"Estonia",
"Botswana",
"Faroe Islands",
"Algeria",
"Liberia",
"Monaco",
"Kazakhstan",
"Congo, The Democratic Republic of the",
"Cook Islands",
"Zimbabwe",
"Mauritius",
"Bonaire, Sint Eustatius and Saba",
"Belarus",
"Tonga",
"Western Sahara",
"Côte d'Ivoire",
"Bermuda",
"Latvia",
"Lithuania",
"Cocos (Keeling) Islands",
"Marshall Islands",
"Antigua and Barbuda",
"North Macedonia",
"Puerto Rico",
"Aruba",
"Antarctica",
"Andorra",
"British Indian Ocean Territory",
"Montenegro",
"French Southern Territories",
"Burundi",
"Cuba",
"Mozambique",
"Niger",
"Iran, Islamic Republic of",
"Réunion",
"Sri Lanka",
"Gabon",
"Morocco",
"Pakistan",
"Tunisia",
"Isle of Man",
"Maldives",
"Montserrat",
"Uganda",
"Curaçao",
"Nigeria",
"Turkmenistan",
"Cayman Islands",
"Mayotte",
"Benin",
"Kiribati",
"Bhutan",
"Ethiopia",
"Micronesia, Federated States of",
"Madagascar",
"Guam",
"Nauru",
"Dominican Republic",
"Jersey",
"French Polynesia",
"South Georgia and the South Sandwich Islands",
"New Caledonia",
"Comoros",
"Falkland Islands (Malvinas)",
"American Samoa",
"Barbados",
"Timor-Leste",
"Guyana",
"Dominica",
"Greenland",
"Cambodia",
"Zambia",
"Albania",
"Kyrgyzstan",
"Trinidad and Tobago",
"Burkina Faso",
"Djibouti",
"San Marino",
"Georgia",
"Guadeloupe",
"Kenya",
"Chad",
"Guinea",
"Tokelau",
"Angola",
"Armenia",
"Suriname",
"Togo",
"Pitcairn",
"Svalbard and Jan Mayen",
"Saint Lucia",
"Central African Republic",
"Eritrea",
"Jordan",
"Tanzania, United Republic of",
"Bouvet Island",
"Saint Pierre and Miquelon",
"Haiti",
"Bangladesh",
"Papua New Guinea",
"Azerbaijan",
"Tuvalu",
"Guinea-Bissau",
"Cabo Verde",
"Wallis and Futuna",
"Martinique",
"Virgin Islands, British",
"Ghana",
"Eswatini",
"Åland Islands",
"Seychelles",
"Niue",
"Rwanda",
"Sint Maarten (Dutch part)",
"Tajikistan",
"Vanuatu",
"Norfolk Island",
"Grenada",
"Cameroon",
"Anguilla",
"Namibia",
"Samoa",
"Solomon Islands",
"Saint Kitts and Nevis",
"Lao People's Democratic Republic",
"Sudan",
"South Sudan",
"Palau",
"Lesotho",
"Belize",
"Saint Barthélemy",
"Malawi",
"Mali",
"Saint Martin (French part)",
"Sao Tome and Principe",
"Gambia",
"French Guiana",
"Equatorial Guinea",
"Myanmar",
"Mauritania",
  
]

pop_2023 = pop_2023[pop_2023["country"].isin(countries_in_data)]

# Save clean CSV for SQL import
pop_2023.to_csv(POP_CLEAN_CSV, index=False)
print("Population data cleaned and saved to:", POP_CLEAN_CSV)
